// src/models/Sport.ts
export enum Sport {
  None = "None",
  Soccer = "Soccer",
  Basketball = "Basketball",
  Hockey = "Hockey",
  Cricket = "Cricket",
}

export enum League {
  None = "None",
  EPL = "English Premier League",
  LaLiga = "La Liga",
  SerieA = "Serie A",
  Bundesliga = "Bundesliga",
  NBA = "National Basketball Association",
  Euroleague = "Euroleague",
  NHL = "National Hockey League",
  ICL = "Indian Cricket League",
}